Fixed service reauth timeout
In the previous version, on every service reauth message we had a
timeout meaning "Reply me back in at most X seconds". The idea was that
the service would send another reauth in exactly X seconds.

The FsmService state machine was actually waiting for more than this,
and the client considering the service didn't reply on time