Rearrange disconnect and state transition
This avoids the state machine being in the incorrect state in the
case where the 'disconnected' notification happens immediately
(that is, it happens *inside* the disconnect function call, without
entering the event loop).

If the FSM is in the wrong state, an error will be triggered, rather
than the intended reconnect. This change ensures the FSM will be in
the right state, by setting the state up before the disconnect is
requested.