Switch from including malloc.h to stdlib.h
The malloc.h header file is apparently Linux specific, and
stdlib.h is the correct header file to use for the standard
heap functions.

In addition, the include ordering has been changed slightly to
avoid having to typedef Buffer in channel.h. This caused it to
get typedef-ed twice, which gcc on macos wasn't happy about.
The change is good anyway, since this shouldn't have been
necessary.