Increase sync resolution; remove semaphore exclusivity
The O_EXCL flag passed when creating the semaphores was causing
execution to fail on macOC. I've therefore removed exclusivity,
which looks horrible to me, but allows execution to continue.

Similarly, synchronisation was failing with event steps of 100,
resulting in the FSM-Service test failing with a timeout error.
Increasing the timer resolution to steps of 25 allows all the
tests to pass successfully.